Added crypto functionality

+++ server.js
@@ -21,6 +21,7 @@ const mongoose = require('mongoose');
 mongoose.connect(process.env.DATABASE_URL);
 const NFT = require('./models/nft');
 const DEV = require('./models/dev');
+const Crypto = require('./models/crypto');
 const db = mongoose.connection;
 db.on('error', console.error.bind(console, 'connection error:'));
 db.once('open', function () {
@@ -37,7 +38,13 @@ app.delete('/nft/:id', handleDeleteNft);
 app.put('/nft/:id', handleUpdateNft);
 app.put('/dev/:id', handleUpdateDev);
 
-// functions for Paths
+// Crypto Path
+
+app.get('/crypto', handleGetUserCrypto);
+app.post('/crypto', handleCreateCrypto);
+app.delete('/crypto/:id', handleDeleteCrypto);
+
+// functions for NFT
 
 async function handelgetAllDevs(req, res) {
   try {
@@ -87,11 +94,14 @@ async function handleCreateNft(req, res) {
 }
 
 async function handleDeleteNft(request, response, next) {
+  const { id } = req.params;
   try {
-    await NFT.findByIdAndDelete({
-      _id: request.params.id,
-      email: req.user.email,
-    });
+    const coin = await NFT.findOne({ _id: id, email: req.user.email })
+    if (coin) {
+      await NFT.findByIdAndDelete({
+        _id: id
+      });
+    }
     response.status('200').send('NFT was deleted');
   } catch (error) {
     next(error.message);
@@ -111,6 +121,8 @@ async function handleUpdateNft(req, res) {
   }
 }
 
+//Functions for Dev
+
 async function handleUpdateDev(req, res) {
   try {
     const updateDev = await DEV.findByIdAndUpdate(req.params.id, req.body, {
@@ -122,12 +134,60 @@ async function handleUpdateDev(req, res) {
   }
 }
 
+//Functions for Crypto
+
+async function handleGetUserCrypto(req, res) {
+  try {
+    const coins = await Crypto.find({ email: req.user.email });
+    res.status(200).send(coins);
+  } catch (error) {
+    res.status(400).send("Could not find coins");
+  }
+}
+
+async function handleCreateCrypto(req, res) {
+  try {
+    const coins = await Crypto.create({ ...req.body, email: req.user.email });
+    res.status(204).send("coins were succussfully added");
+  } catch (error) {
+    res.status(400).send(error.message);
+  }
+}
+
+async function handleDeleteCrypto(req, res) {
+  const { id } = req.params;
+  try {
+    const coins = await Crypto.find({ _id: id, email: req.user.email });
+    console.log(coins);
+    if (coins) {
+      await Reading.findByIdAndDelete(id);
+      res.status(204).send("coins were succussfully deleted");
+    }
+  } catch (error) {
+    res.status(400).send(error.message);
+  }
+}
+
+async function handleCreateCrypto(req, res) {
+  try {
+    const coins = await Crypto.create({ ...req.body, email: req.user.email });
+    res.status(204).send("coins were succussfully added");
+  } catch (error) {
+    res.status(400).send(error.message);
+  }
+}
+
 //Landing page for testing purposes
 app.get('/', (request, response) => {
   response.send('We Are Working!!!');
 });
 
+app.get('*', (req, res) => {
+  res.send("Page not found");
+})
+
 //Error Handling
+
 app.use((error, req, res, next) => {
   res.status(500).send(error.message);
 });
